Refactor DTOs and enhance Swagger documentation
Consolidated and renamed DTO files for appointments and staff to use a unified 'req.dto.ts' naming convention. Added @ApiProperty decorators to DTO fields for improved Swagger documentation. Updated controller imports and removed the now-redundant fetch-staff.dto.ts. Adjusted Swagger config to use API key authentication and improved the API description. Expanded and modernized the README with detailed setup, usage, and API documentation.
