+++ src/staff/dto/req.dto.ts
@@ -1,4 +1,6 @@
-import { Type } from 'class-transformer';
+import { PaginationDto } from '@common/utils/pagination';
+import { ApiProperty } from '@nestjs/swagger';
+import { Transform, Type } from 'class-transformer';
 import {
   IsArray,
   IsInt,
@@ -13,15 +15,18 @@ import {
 } from 'class-validator';
 
 export class CreateStaffDto {
+  @ApiProperty()
   @IsString()
   @IsNotEmpty()
   name: string;
 
+  @ApiProperty()
   @IsString()
   @IsOptional()
   @IsTimeZone()
   timezone: string = 'UTC';
 
+  @ApiProperty()
   @IsArray()
   @ValidateNested({ each: true })
   @Type(() => CreateWorkingHoursDto)
@@ -30,20 +35,33 @@ export class CreateStaffDto {
 }
 
 export class CreateWorkingHoursDto {
+  @ApiProperty()
   @IsInt()
   @Min(0)
   @Max(6)
   dayOfWeek: number;
 
+  @ApiProperty()
   @IsString()
   @Matches(/^([01]\d|2[0-3]):([0-5]\d)$/, {
     message: 'startTime must be in HH:mm format',
   })
   startTime: string;
 
+  @ApiProperty()
   @IsString()
   @Matches(/^([01]\d|2[0-3]):([0-5]\d)$/, {
     message: 'endTime must be in HH:mm format',
   })
   endTime: string;
+}
+
+export class FetchStaffDto extends PaginationDto {
+  @ApiProperty()
+  @Transform(({ value }) => (value ? Number(value) : 1))
+  page: number = 1;
+
+  @ApiProperty()
+  @Transform(({ value }) => (value ? Number(value) : 10))
+  size: number = 10;
 }
